requirements.txt
chardet==5.2.0
evaluate==0.4.1
fastapi==0.104.1
langchain==0.0.340
langserve==0.0.30
numpy==1.26.2
pandas==2.1.3
pydantic==1.10.7
python-dotenv==1.0.0
scikit_learn==1.3.2
torch==2.1.1
transformers==4.35.2
uvicorn==0.24.0.post1
